DocBook documentation: complete rewrite of the C chapter
The C chapter of our documentation was completely out of date so had
to be rewritten.  The new writeup emphasizes C99 standards compliance,
provides a list (which will eventually be referred to from elsewhere
in the documentation) of well-documented public PLplot C types
including the relevant callback types, and concludes with a paragraph
emphasizing our C examples as prototypes of how to call functions in
our core C library and the traditional (Makefile + pkg-config) and
CMake-based build systems for our installed C (and other language)
examples as prototypes for how to build C applications that use the
PLplot C library.

N.B. I suggest that all our different language chapters should be
changed to include such a paragraph if they don't do that already.