libplplot: logic improvements for pl_create_tempfile and pl_create_tempfifo
Instead of using duplicate code to create the list of directories to
attempt to use for temporary file and fifo creation,
implement a static function, get_tmpdir_list, to create the list of
possible directories.

The list of directories that is tried is changed to be as follows:

The directory names in the environment variables:
"TMP" [Windows only],
"TEMP" [Windows only],
"TMPDIR" [Unix only];
the value of the P_tmpdir macro as defined in <stdio.h> [Unix only];
the current working directory; and
the following hard-coded locations as a last resort:
c:\\windows\\Temp [Windows only], and
"/tmp" [Unix only].

The above choices make the following changes in the directory list from
the previous code.  "TMP" is added for the Windows case; the misspelling
P_TMPDIR ==> P_tmpdir is corrected and P_tmpdir is now only used for
the Unix case; and the current working directory is
added.

Note get_tmpdir_list delivers a list of non-NULL pointers that point
to NULL-terminated character strings, and it is up to
pl_create_tempfile and pl_create_tempfifo to use this list one string
at a time to find and use the first string that is a suitable
directory location for creating a temporary file or fifo.  This is a
substantial improvement on the old code which included NULL pointers
in the list and which only tried the first list item that corresponded
to a non-NULL pointer.

A pldebug call is now made from both pl_create_tempfile and
pl_create_tempfifo to make it possible to debug cases where
there are difficulties in generating temporary files or fifos.

The bugs that were fixed in the previous code were the following:

* The misspelled P_TMPDIR macro mentioned above.

* Errors in the open function flags that are part of the solution when
mkstemp is not available. The O_BINARY, _O_SHORT_LIVED, and
_O_TEMPORARY flags are now removed for the Unix case since those flags
are not defined for that case, but those flags continue to be used for
the Windows case since they are available for that platform.

* A check for errors was implemented for the mkdtemp call in
pl_create_tempfifo.  Without such a check the old code segfaulted when
an attempt was made to create a fifo in an inappropriate directory in
the list of possibilites.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the test_c_tk, and svg targets (to
satisfy necessary dependencies) and then running both

valgrind examples/c/x00c -dev svg -o test.svg -debug >| debug.out 2>&1

and

valgrind examples/c/x00c -dev tk -debug >| debug.out 2>&1

(The tk case is the only one that currently tests
pl_create_tempfifo.)
For the svg case, a perfect valgrind report was obtained (0 errors, no leaks are
possible) while for the tk case the valgrind report was good (0
errors, but leaks are possible).  I ascribe that "leaks are possible"
result to issues with libtcl and libtk.  In both cases the
appropriate debug messages were obtained as well as the appropriate
plots.

I repeated the same two tests for the 3 remaining possible combinations of the
following two local changes to the code:

* replace the preprocessing checks of the PL_HAVE_MKSTEMP macro with
PL_HAVE_MKSTEMP1 to emulate the case where mkstemp is not available.

* replace some or all of the valid directory possibilities with "whatever" to emulate
when some or all of the directories on the list are inappropriate.

The same good valgrind report results were obtained for all these three cases
which validates how these cases are currently handled by the
pl_create_tempfile and pl_create_tempfifo code.

In sum, from all this testing on a Unix platform I am pretty sure this
code will work on most Unix platforms, but testing is
required for the Windows case to make sure I haven't introduced
any showstopping bugs for that case.