Update Ada docs: Shapefile-related overloads
Adds a paragraph in the "One-offs" section of the Ada documention describing the overloaded procedures for plmapfill and plmapline that were required to allow the Ada programmer to code according to the PLplot documentation which allows passing two types of entities in the plotentries slot, an array of integers or a null pointer. Also mentions the the additional procedures  which do not require conjuring a null pointer.