Change Ada example names on web site buttons
This trivial commit changes the names on the web site's example pages buttons from Ada(thick) and Ada(thin) which are misleading because both bindings are thick, to Ada(Standard) and Ada(Traditional) to indicate more clearly the reasons for the two bindings—Traditional uses normal, documented subprogram names and Standard uses names that are thought to appeal to Ada programmers. This is in anticipation of changing the Ada example file names/main program names and Ada binding names accordingly.

Once the example file names are actually changed, lines 185 and 186 of examples.php will have to be changed.

This change is not tested but should be checked after the next web site build.