Update parity_bit_check.sh and associated exclude file
As an extremely incomplete but still interesting check for file
corruption, the scripts/parity_bit_check.sh script searches for any
ascii characters that have their eighth bit set for all the files in
our source tree that are still considered to be pure ascii.  Files to
be ignored by this script (i.e., binary files or files with non-ascii
UTF-8 characters) are indicated in scripts/parity_bit_check.exclude using
regular expressions.

The updates to the script were to use git rather than svn to do the
equivalent of svn export using git --checkout-index. The updates to
the exclude file contain changes to the regular expressions for the
Ada examples with UTF-8 characters consistent with the recent changes
to the Ada example naming conventions.  Additional file patterns also
had to be added to the exclude file to account for the additional
files with non-ascii UTF-8 characters that are present now compared to
when this script was last invoked several years ago.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by building the parity_bit_check executable (using
make parity_bit_check), putting the absolute location of the utils
subdirectory of the build tree on the PATH, and running
scripts/parity_bit_check.sh.  The result was no output, i.e., every
file checked had no ascii parity bit set. I also double-checked (by
locally modifying the script to not remove
/tmp/temporary_source_tree_list and then looking at the contents of
that file) that git --checkout-index was working properly to export
the entire working directory of the local plplot git repository. It
turns out there is a bug in git --checkout-index (see
<https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=823384>) such that
it quietly does not export the entire source tree if invoked from
within a subdirectory of the source tree, but I worked around that
issue by using cd appropriately in the script.