DocBook documentation: Remove OS notes chapter
This removed chapter (which gave instructions concerning how to build
PLplot for various platforms) was completely out of date/misleading
compared to our wiki which should be the primary source of this
information. Also, the only internal reference to this chapter was in
the introduction, and that part of the introduction has now been
implemented a different way by providing generic build instructions
with reference to the wiki for the details.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by using the cmake -DBUILD_DOC=ON option, and by
building the "html" target without validation or title-page rendering
issues.