DocBook documentation: Improve error messages from XML::DOM::Parser in perl scripts
The improvement in each perl script was to make the following change:
-$p = new XML::DOM::Parser;
+$p = new XML::DOM::Parser(ErrorContext => 2);
 $sects = $p->parse ($api)->getElementsByTagName ("sect1");

The effect of this change is to add some context to error messages
generated when parsing the temporary file $api. This newly introduced
context for error messages makes it much more convenient to discover
the source of errors in the actual files plplotdoc_info.xml and
api.xml which are concatanated together (along with some relevant
entity definitions) to create $api.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian jessie) by using the cmake -DBUILD_DOC=ON option and by
building the "man" target.  As a result of the above change, the perl
script error created by using the "le" entity during local editing
of api.xml was quickly spotted and addressed by replacing "le" by
"leq" (one of the entities specifically defined in the perl scripts)
for my locally updated api.xml.  The plan is to commit that
locally updated (with "le --> leq" correction for the update) api.xml as
soon as I am finished making changes to it.