Changes required to support the NAG Fortran compiler
The NAG Fortran compiler is much more strict in checking the code against the prevailing standard than most other compilers. Our work to get the Fortran binding and examples accepted by this compiler revealed a number of deivations from the standard were present in the code. They mostly have to do with the finer details of the Fortran/C interfacing.

Some more work needs to be done for the Windows platform (MinGW-w64/MSYS2) to make it work smoothly, because the support by CMake for this compiler is not yet fully developed and the installation of the NAG Fortran compiler on Windows uses spaces in one of the directory names, which is something our build system does not handle well.

Tested by Arjen Markus (arjenmarkus@sourceforge.net) on MinGW-w64/MSYS2 and Linux