Adjust Fortran binding to remove unused arguments and variables
The NAG Fortran compiler produced a number of issues that were related to unused arguments and variables. One set of these messages concerned the argument "defined" for plshade and similar routines. This argument has been removed from the API. While this introduces a limited incompatibility it simplifies the interface, especially as the argument was dummy for all current specific interfaces. The examples have been adjusted to reflect this change.

Some other unused arguments exist because the routines are callbacks and therefore the API is defined at the C side.

The various examples turned out to have a small number of similar issues, nothing of any consequence though, but to make them as clean as possible, several changes were made.

Tested by: Arjen Markus (arjenmarkus@users.sourceforge.net) on Cygwin and MS Windows with the Intel Fortran compiler