Make the Tcl examples use the redacted API as far as possible
The examples have been adapted to use the readacted API as far as possible. In some cases a smaller part of the vector of data is used, so that there the number of data points is specified explictly.

This change required a change in some C routines to actually take care of the reduced API. (We currently support a range of argument lists for some of these commands.)

Tested by Arjen Markus (arjenmarkus@users.sourceforge.net) on Cygwin