Check the data argument for callbacks that do not use them
The purpose of checking the data argument for the two Fortran callback routines pltransformf2c and pllabelerf2c is twofold: make sure the binding can be compiled with as little warnings as possible and to make sure that the correct version of the routine is being used.

The most critical compiler, NAG Fortran, could not be used unfortunately (expired licence), but earlier experiments indicated this change makes the warnings go away.

Tested by Arjen Markus (arjenmarkus@users.sourceforge.net) on Windows+Cygwin