Implement new test_fortran project to do what its name implies
This project was inspired by the existing test_ada project.  Similarly
to that project but for Fortran rather than Ada it creates a library
with a routine that emits the "hello, world" phrase, and an executable
that calls that routine.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) with gfortran by following all the steps in
cmake/test_fortran/README.