test_ada project: For Windows case properly support build-tree tests.
This is done by collecting build-tree dll's in the dll subdirectory of
the top-level build-tree directory, and it is the user's
responsibility to put that dll subdirectory on their PATH for the
build-tree case, but otherwise (install-tree case) ignore that
subdirectory.  Note, such PATH manipulations have already been set up
in cmake/test_ada/scripts/comprehensive_test.sh so this fix should
"just work" when someone tests the test_ada project on Cygwin and/or
MinGW-w64/MSYS2 without any need for them to do PATH manipulations
(unless they run tests by hand).

Untested since this added logic has already been proved to not
interfere with the test_fortran Linux case, and actual test of
test_ada on Windows platforms (which will thoroughly test this logic)
isn't planned any time soon (since it is already known that Ada
language support does not work correctly for the Windows case and
debugging that case with test_ada on Cygwin and/or MinGW-w64/MSYS2 is
going to take a major effort).