Made backward compatibility for PLpointer in Ada
A recent change in the Ada bindings changed PLpointer to PL_Pointer, causing a backward incompatibility. See fdd31ce6. The present change restores that backward compatibility but marks the PLpointer spelling as deprecated.

Tested by temporarily changing all instances of PL_Pointer to PLpointer in both the traditional and standard versions of Example 14.