Restore backward compatiblity for Ada bindings
Restore backward compatibility for the "standard" Ada binding that was lost when it was renamed from PLplot to PLplot_Standard, with files plplot.adb and plplot.ads renamed to plplot_standard.ads and plplot_standard.adb.

A single new file was added, with the old name, plplot.ads. No .adb file is required. Users that have used the PLplot name need not change their code although a "deprecated" note is added herein.