Build system: Update FindwxWidgets.cmake to be consistent with cmake master branch
The current (v3.6.2-1012-g4171b86) cmake master-branch version of
FindwxWidgets.cmake is being actively maintained and appears to be the
same or superior to our own version in all respects.  So I have
changed our version to be the same as the cmake master-branch version
except for necessary licensing language changes and the different
method we use find_package_handle_standard_args which should work for
whatever cmake version >= 3.0.2 that is being used by the user.

So the net effect of this commit is users of older cmake versions such
as 3.0.2 will have access to the latest official cmake version of
FindwxWidgets.cmake which contains many bug fixes compared to the
3.0.2 version.  And that desired result should continue if we are
careful to update our version of this find module to the latest
official cmake version whenever there is a significant change in that
version.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by running cmake and building the wxwidgets device.
Testing on additional platforms is requested.