Build system: filter source tree and build tree -I flags from pkg-config files
Such -I flags do occur, e.g., for the tk device driver when it is
embedded in libplplot when -DENABLE_DYNDRIVERS=OFF, but I cannot
imagine any case where the resulting pkg-config file should be
generating such -I flags for applications and libraries which link to
the installed PLplot libraries.  In other words, the install tree is
supposed to work when the source tree and build tree have been
completely removed, and the previous -I flags in the pkg-config files
that referred to source-tree and/or build-tree locations violated this
philosophy.  Thus, our build system now filters these -I flags out
of the generated pkg-config files as a matter of principle.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> as part of the
successful comprehensive testing of the last commit which was limited
to just the Tcl/Tk components of PLplot.  So the component testing was
limited, but did include running cmake to configure at least two *.pc
files without issues, and running and testing the traditional build of
the installed examples occurred without running into issues so the
conclusion is there is unlikely to be any typographical error in the
changes of this commit (which simply added the required filtering
commands to the CMake function used to configure the *.pc files).
Therefore, it is very likely this limited component testing is more
than enough for the reasons mentioned in the first paragraph, but
later on, as a matter of course, more complete component testing will
be done, and if any issues are found, this commit is separated from
other changes so should be easy to find with git bisect.