libtclmatrix: fix bugs in matrix max/min and reimplement matrix initialization and assignment
The last libtclmatrix commit introduced bugs in the matrix max min
C logic (such that a default length always lead to an error).  These
bugs have been fixed in this commit.

Note that for the previous matrix initialize code, there were severe
memory management issues if the RHS had more numerical items than fit
into a column of the LHS matrix, and for other kinds of shape
mismatches silent failures occurred.  In other words, the only way the
old code worked was if the size and shape of the lists on the RHS was
exactly equivalent to the size and shape of the matrix on the LHS.  So
that precluded, for example using a matrix slice on the RHS (which
gets interpreted as a shallow list) if the LHS matrix is smaller than
that RHS shallow list or if the LHS matrix is multi-dimensional.

Because of these issues, I reimplemented the matrix initialization so
there are no constraints on the shape or size of the matrix
initializer list that appears on the RHS of the "=".  As the result of
this change, the following exotic matrix initialize example works:

pltcl> matrix a f 2 2 3 = {{{{1} {2}}}} 3 4 5 6 7 8 9 10 11 12 13 14
a
pltcl> a dump
1.0 2.0 3.0
4.0 5.0 6.0

7.0 8.0 9.0
10.0 11.0 12.0

Note a less exotic but more useful example demonstrating
resizing/reshaping capability is the following:

pltcl> matrix x f 2 7 = 0 1 2 3 4 5 6 7 8 9 10 11 12 13
x
pltcl> x dump
0.0 1.0 2.0 3.0 4.0 5.0 6.0
7.0 8.0 9.0 10.0 11.0 12.0 13.0

pltcl> matrix a f 2 2 3 = [x 0 :]
a
pltcl> a dump
0.0 1.0 2.0
3.0 4.0 5.0

6.0 0.0 0.0
0.0 0.0 0.0

Note that the array and lists are processed in the natural order (row
major order as in C for the array or array slice, and in the order of
the list items that are not lists themselves when the RHS is an
explicit list rather than a matrix slice.  Furthermore, if too many
initializers are used on the RHS as in the first example, they are
quietly ignored; while if too few initializers are used on the RHS as
in the second example, then the corresponding LHS matrix elements
are left unchanged from their initial 0 (or 0.) values they are given
before further array initialization is attempted.

Note array assignment had substantial issues as well.  (Only one
number used from the RHS of the "=" to initialize the entire
specified slice of the matrix on the LHS.) So I reimplemented
that as well so that all forms of the RHS of the "=" above can
also be used on the RHS of the "=" of an array slice assignment, e.g.,

pltcl> matrix x f 2 7 = 0 1 2 3 4 5 6 7 8 9 10 11 12 13
x
pltcl> matrix a f 2 2 3
a
pltcl> a : : : = [x 1 :]
pltcl> a dump
7.0 8.0 9.0
10.0 11.0 12.0

13.0 0.0 0.0
0.0 0.0 0.0

The new rule for matrix assignment is if the RHS is a pure single
number that is completely valid according to strtod (for the
floating-point matrix case) or strtol (for the integer matrix case),
then all matrix elements on the LHS are updated to that number, but
otherwise the same routine is called to assign values to the elements
of the LHS matrix slice as is used for matrix initialization with
elements outside the LHS slice unaffected.  Furthermore, if too many
values are specified on the RHS, they are quietly ignored; while if
too few values are specified on the RHS then the corresponding LHS
matrix slice elements are left unchanged from their previous values.

As a test of this new matrix initialization and assignment code, I
have added some matrix initialization and assignment examples to the
already existing slice examples in bindings/tcl/test_tclmatrix.tcl.
This file is used as the source of all tclmatrix examples tested by
the check_libtclmatrix_capabilities target, and I have adjusted the
comparison output for that target, bindings/tcl/test_tclmatrix.out, to
be consistent with these additional test examples.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by building the test_interactive, test_noninteractive,
and check_libtclmatrix_capabilities targets with no obvious build or
run-time issues. (The test_interactive target originally found the
matrix max/min bug which has been fixed in this commit.) I also
sourced bindings/tcl/test_tclmatrix.tcl under bindings/tcl/tclsh when
run with valgrind, and there were no memory management issues reported
by valgrind. Finally, I have tried the examples specified above under
valgrind and obtained the stated results with no memory management
issues being reported by valgrind.