plplottcltk library: comment, clean up, and reorganize Tcl code used to generate the interface
The added comments are by no means complete and were simply added as I
finally understood the purpose of various bits of the Tcl code.  The
cleanup was to remove some set commands for variables which are not
used.  The reorganization was to replace the "round" loop by one call
(if nredacted == 0) or two or three calls (if nredacted > 0) to a new
process_args procedure which does most of the work that formerly
resided in the "round" loop.  This reorganization includes optional
logic for removing the generation of the non-redacted API for Tcl, but
that code has not been exercised yet (because if_non_redacted is set
to 1).

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by building the plplottcltk library and confirming
that the three source code files generated by the updated pltclgen.tcl
(tclgen.c, tclgen.h, and tclgen_s.h in bindings/tcl in the build tree)
are identical to what was produced by the previous version of
pltclgen.tcl.