test_automoc project: Add the helloworld_automoc_same_include2 target
This additional target is similar to both the
helloworld_automoc_same_include and helloworld_automoc_same_include1
targets. The implementation codes for the three targets are all
identical and contain both

(and at the bottom of the file)

The idea is for all three targets to generate name collisions for the
generated moc_test_q_object.cpp file in the same build tree
subdirectory to see whether CMake-3.7.0 diagnoses this issue.  The
helloworld_automoc_same_include1 target uses the identical source code
in the same directory as the helloworld_automoc_same_include target,
while the helloworld_automoc_same_include2 target does the same thing
for source code that is copied from the directory used by the first
two targets to a different directory.  I thought that directory
difference might force cmake-3.7.0 to generate the name collision
diagnostic that is mentioned in the documentation, but no such
diagnostic was generated for any of these cases.