test_linux_ipc: Finish this test project
I updated both the pshm_write and pshm_read applications to
the "Unnamed semaphores example" paradigm mentioned in the
README, and also implemented a pshm_unlink application to deal
with cleanup of the shared memory if psh_write (which creates
that memory) must be interrupted for some reason.

Tested by: Alan W. Irwin <airwin.users.sourceforge.net> on Linux
(Debian testing) following the directions in the README.  The result
was a big efficiency success; a file copy using the combination of
pshm_write and pshm_read only required 0.3 seconds to complete for
a 25MB file.