Build system: reduce loudness of warnings when a compiler is broken or missing
Also, for this case advertise (and implement) the
-DENABLE_compiler_diagnostics=ON option to give full CMake diagnostics
when a compiler is broken or missing.  This option should be useful
to users where the broken/missing compiler is a surprise.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by deliberately breaking the gfortran compiler
by specifying

export FFLAGS=-whatever

and checking the messages that result from
-DENABLE_compiler_diagnostics=ON or (the default)
-DENABLE_compiler_diagnostics=OFF.