Implement submission of dashboards to the my.cdash.org cdash server
The new CTestConfig.cmake file was provided by my.cdash.org to
facilitate submitting dashboards (defined to be essentially the complete
set of information associated with a particular ctest run) to the new
"PLplot_git" project I created at my.cdash.org. Note the "ctest -D
Experimental " command I used below to test this project, can be
executed by anybody to create a dashboard and send it to the
"PLplot_git" project.  However, since this is a free service (provided
by Kitware) there is a limit of a total of 10 dashboard submissions
per day.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by running the following commands:

cmake -DBUILD_TEST=ON ....
make -j4 all  #N.B. the all target must always be built before ctest is run
ctest -j4 -D Experimental

(The -j4 option for makes builds the all target in parallel, and the
-j4 option for ctest runs the tests in parallel.)

That ctest command (which required ~5 minutes) executed 31 tests that
are enabled by the above cmake command; created a dashboard of those
results; and submitted that dashboard to the above project. That
dashboard (and one previous one I did) can be viewed at
<http://my.cdash.org/index.php?project=PLplot_git> and shows the whole
process was a complete success.  Note that virtually every field in a
given dashboard is presented on the my.cdash.org cdash server as a
clickable link that gives more information about all aspects of the
dashboard in an extremely convenient format.