Build system: Fix stale cached variable bug in determining correct HAS_LIBQHULL_INCLUDE value
If HAS_LIBQHULL_INCLUDE was ON as a result of the first cmake run, for any subsequent
run the cached variable QHULL_INCLUDE_DIR would be true and HAS_LIBQHULL_INCLUDE would
(incorrectly) be set to OFF as a result.  The fix was to look for both
forms of header with one command and store the result in the cached
variable QHULL_INCLUDE_FILE, and then parse that cached result to see if
it was the libqhull form or not.  Subsequent runs would skip the find
(because the cached value would show that had already been done) and
continue to parse that result the same way.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux by
running cmake twice and looking at the QHULL-related results for the
case when the libqhull form of header was present and also the case
when it was not.  In all cases got the expected results for
HAS_LIBQHULL_INCLUDE which were consistent between cmake runs.