Remove plplot_test/test_single_python.sh script
This bash script historically was used to run a single Python example.  It
had fallen into disrepair (it would not work properly on standard
examples 14, 17 and 31, and also it needed a fixup to work with the
namespaced version of the xw?? modules we have now), and was located
in an awkward directory in the source tree from which it would have
had to be run from examples/python in the build tree.  Also this
script has been completely superseded by the examples/python/x??
Python scripts and is referred to by no other script or file (except
for historical data) within our source tree. Therefore, deep-six this
script.