Python + Plframe/Tk/Tcl GUI: large update
Changed pytkdemo (which demonstrates this GUI) to add examples 0, 9,
and 18 to the GUI and change the standard example modules that were
imported from the x??.py versions to the xw??.py versions.  The x??.py
versions had historically served as the starting basis for the xw??.py
versions, but the latter had been greatly expanded, modernized, and
maintained while the former were essentially unmaintained.  Thus, the
only reason to keep the x??.py versions around was because the xw??.py
versions had originally been changed to using no plplot namespace (a
historical error by yours truly because I didn't understand Python
very well way back when) so could not be called from pytkdemo.  That
historical error has now been rectified allowing the present change to
pytkdemo so it now imports the xw??.py versions of the standard
example modules.

The modernized/expanded standard example modules required large
changes to Plframe.py to support the additional PLplot API used in
those examples including all the PLplot constants and a substantial
number of PLplot functions which had not been supported before.

N.B. pytkdemo currently does not display examples 14, 15, 17, and
19-33. If that demo was modified to add in some or all of those
missing examples more PLplot functions would have to be added to
Plframe.py.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by building the test_pytkdemo target and
clicking on each of the subset of examples that are enabled
in that GUI (examples 0 through 18 excluding 14, 15, and 17).

Those examples looked find with no observable cross-talk between them
except for one intermittent colour issue which is when I clicked on
example 0 through 3 in that exact order, example 3 turned up with the
wrong colours.  But I could not reproduce the problem reliably so I
suspect these intermittent colour issues are the result of some subtle
memory management issue.

In sum, pytkdemo now has a new lease on life with very few issues
compared to what it had before with the x??.py version of the
standard example modules.  And this commit prepares the way
for removing the x??.py version of the standard examples in
the next commit.