Build system: Fix linking issues
For the plplot_widgetmodule shared object used by pytkdemo, the ENABLE_DYNDRIVERS
OFF linking case only works if -dev tk code is incorporated into libplplot.
So add some "AND (ENABLE_DYNDRIVERS OR PLD_tk)" Boolean logic to only
build the plplot_widgetmodule shared object and the test_pytkdemo target
that depends on it only for this (correct) case.

Add list of external devices to interactive devices to determine
<driver>_targets in test system targets that are enabled in the
examples subdirectory.  This change makes sure that qt_targets and
cairo_targets are correctly calculated for the case where only the
extqt form and not the interactive form of the qt or cairo devices
is enabaled.

Also, I took this opportunity to allow test_tk and test_interactive
general interactive targets to depend on test_pytkdemo since that GUI
has become so reliable now.