Comprehensive test script: drop dllplplot_stubs.so from the ldd -r reports
We deliberately drop dllplplot_stubs.so from the ldd -r reports on
Linux because there is no known way to set the rpath for that
ocamlmklib-generated shared object at the moment.  See plplot-devel
mailing discusion on 2016-12-09 with the subject line
"Setting rpath for dllplplot_stubs.so appears not to be possible".