Build system bug fix: Set USINGDLL macro for automoc-generated source code
The USINGDLL macro must be set in the shared libraries case in
order for visibility of symbols to be implemented correctly.

Our previous method of specifying this macro set the the source-code
property COMPILE_FLAGS (or at least one part of that space-separated
list) to "-DUSINGDLL".  But that method fails on automoc-generated
source code (and possibly other generated source code).  The answer is
to set the target property COMPILE_DEFINITIONS to USINGDLL (which
applies to all source code for the target whether generated or not).
This fix (and also cleanup in style) is an intrusive build-system
change where for libraries this logic is now all centralized with the
set_library_properties function.  As a result of this fix, all
instances of "DUSINGDLL" are now gone from our build-system logic and
either replaced centrally or on-the-spot for modules and executables
by the above COMPILE_DEFINITIONS target property.

This fix allowed us to drop the previous automoc-related workarounds
for the plplot_pyqt4 and plplot_pyqt5 targets that had been applied
previously which was to regenerate automoc results for those targets.
Those workarounds worked on Linux but were causing undefined symbol
problems on Cygwin (and possibly other platforms).  So it is
anticipated this fix will solve that Cygwin issue.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux
(Debian Jessie) by running

scripts/comprehensive_test.sh

The above build system change was intrusive, and a lot of previous
commits were only partially tested.  So note there are no constraining
options used for this script run, i.e., all PLplot components and all
major configurations were tested by it for both the interactive and
non-interactive cases.  The result was near perfect with no obvious
configure-, build-, or run-time issues.  The only exception to this is
noted and discussed below. Furthermore, there were no PostScript
regressions (i.e., the PostScript difference report was perfect other
than the long-standing ocaml issues) and no "ldd -r" issues.

The only issue for this test was
that at one point (after the test_c_ntk target started building, but
that is not necessarily the one causing the issue because of the -j4
jumbling of results) PLplot generated ~5

PLPLOT ERROR, ABORTING OPERATION

messages from a C (or C++) call to

plabort( "<command_name>: Please call plinit first" )

for several different command_names that were executed for the
interactive testing case and the static build.  But without an actual
error exit, and a subsequent run of the appropriate test_interactive
target could not replicate the issue (perhaps because I hit GUI
buttons in a different order?  As a result it is extremely difficult
to find the source of this issue.

The above "PLPLOT ERROR" string is looked for and a specific error
exit is taken for most of the results generated in plplot_test except
for

plplot_test/test_octave_interactive.sh.in

But none of those interactive octave examples are currently exercised.
So I conclude this issue occurred for a different type of interactive
test target that is not coverered by plplot_test.  But since I could
not identify the source of the issue I am documenting it here (in case
I see it ever again) and moving on.